# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: api
  version: 1.0.0
  description: Shawrome Car Sales Platform API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum: [admin, seller, client]
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        phone:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    UsersListResponse:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
    SellerStats:
      type: object
      required: [userId, userName, totalSales, totalRevenue, totalProfit, carsManaged, priceDrops]
      properties:
        userId:
          type: integer
        userName:
          type: string
        totalSales:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
        carsManaged:
          type: integer
        priceDrops:
          type: integer

    # ──────────────────────────── Car ────────────────────────────
...
    Sale:
      type: object
      required: [id, carId, carBrand, carModel, carYear, finalPrice, purchasePrice, profit, profitPercent, sellerId, sellerName, soldAt]
      properties:
        id:
          type: integer
        carId:
          type: integer
        carBrand:
          type: string
        carModel:
          type: string
        carYear:
          type: integer
        finalPrice:
          type: number
        purchasePrice:
          type: number
        profit:
          type: number
        profitPercent:
          type: number
        sellerId:
          type: integer
        sellerName:
          type: string
        soldAt:
          type: string
          format: date-time
    SalesListResponse:
      type: object
      required: [sales, total]
      properties:
        sales:
          type: array
          items:
            $ref: '#/components/schemas/Sale'
        total:
          type: integer
    SalePeriod:
      type: string
      enum: [day, week, month, year]
    SaleStats:
      type: object
      required: [period, totalSales, totalRevenue, totalProfit, totalPurchaseCost, profitPercent, chartData]
      properties:
        period:
          $ref: '#/components/schemas/SalePeriod'
        totalSales:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalPurchaseCost:
          type: number
        profitPercent:
          type: number
        chartData:
          type: array
          items:
            $ref: '#/components/schemas/ChartDataPoint'
    ChartDataPoint:
      type: object
      required: [label, revenue, profit, sales]
      properties:
        label:
          type: string
        revenue:
          type: number
        profit:
          type: number
        sales:
          type: integer
    TopCar:
      type: object
      required: [carId, brand, model, year, salesCount, totalRevenue, totalProfit]
      properties:
        carId:
          type: integer
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        salesCount:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number

    # ──────────────────────────── Dashboard ────────────────────────────
...
    ToggleUserActiveRequest:
      type: object
      required: [active]
      properties:
        active:
          type: boolean

    # ──────────────────────────── Common ────────────────────────────
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      security:
        - bearerAuth: []
      summary: List sales with optional period filter
      parameters:
        - name: period
          in: query
          schema:
            $ref: '#/components/schemas/SalePeriod'
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: sellerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Sales list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesListResponse'
  /sales/stats:
    get:
      operationId: getSaleStats
      tags: [sales]
      security:
        - bearerAuth: []
      summary: Get sales statistics with chart data for a period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/SalePeriod'
        - name: sellerId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Sale statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleStats'
  /sales/top-cars:
    get:
      operationId: getTopCars
      tags: [sales]
      security:
        - bearerAuth: []
      summary: Get top 5 best-selling cars for a period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/SalePeriod'
      responses:
        '200':
          description: Top cars
          content:
            application/json:
              schema:
                type: object
                required: [topCars]
                properties:
                  topCars:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopCar'
  /sales/seller-stats:
    get:
      operationId: getSellerStats
      tags: [sales]
      security:
        - bearerAuth: []
      summary: Get stats grouped by seller (admin only)
      responses:
        '200':
          description: Seller statistics
          content:
            application/json:
              schema:
                type: object
                required: [sellers]
                properties:
                  sellers:
                    type: array
                    items:
                      $ref: '#/components/schemas/SellerStats'

  # ──────────────────────────── Dashboard ────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      summary: List users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'

    post:
      operationId: createUser
      tags: [users]
      security:
        - bearerAuth: []
      summary: Create a new seller/admin account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      security:
        - bearerAuth: []
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    put:
      operationId: updateUser
      tags: [users]
      security:
        - bearerAuth: []
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      operationId: deleteUser
      tags: [users]
      security:
        - bearerAuth: []
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /users/{id}/active:
    patch:
      operationId: toggleUserActive
      tags: [users]
      security:
        - bearerAuth: []
      summary: Toggle user active status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleUserActiveRequest'
      responses:
        '200':
          description: Active status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ──────────────────────────── Favorites ────────────────────────────
...