# lib/api-zod/src/generated/types/topCar.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Shawrome Car Sales Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface TopCar {
  carId: number;
  brand: string;
  model: string;
  year: number;
  salesCount: number;
  totalRevenue: number;
  totalProfit: number;
}