# lib/api-zod/src/generated/types/usersListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Shawrome Car Sales Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { User } from './user';

export interface UsersListResponse {
  users: User[];
  total: number;
}